refactor(DEV-11277): move component files
