+++ packages/sdk-drop-in/src/monite-iframe-app-html.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import { MoniteIframeApp } from '@/MoniteIframeApp';
+import { MoniteIframeApp } from '@/apps/MoniteIframeApp.tsx';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
